requirements.txt
requests==2.32.5
fastapi==0.135.1
uvicorn==0.41.0
watchdog==6.0.0
markdown-it-py==4.0.0